Use raw command-line in CompilationDatabaseIndexingVisitor
This function is trying to reconstruct the clang.exe
command-line by removing -c and -o. However, it was using
command.tokenizeCommandLineToEscaped() which will convert
backslash+quote to literal quote.

This bug would only manifest if user had cl.exe flags that
contained had a a literal quote inside of a quoted block.
For example,
clang.exe -c "The \"main\" file.cpp"

Obviously, this is rare which is probably why we never got
a bug report.

Bug: NA
Test: NA

Change-Id: I7c01fca182574e3392355a1ac64477dbb82f7eae