Increase --jobs to 150.
This enables Windows to use more RBE workers when executing
remotely. In rare instances, when Windows has to run a lot of
long-running tests, there won't be enough workers to maximize
test parallelism.

We have not seen our RBE worker pools become too heavily overloaded
with the current job limit, should be OK with this increase.

Bug: none
Test: none
Change-Id: Ica74115c87c43786e657b24c90b31dc9d4cbe43d