Fix ability to use NDK r18.
Apparently using NDK r18 *at all* has been broken since 3.6ish.

https://stackoverflow.com/q/61461145/632035 pointed this out, and
something as trivial as building the hello JNI NDK sample with r18
reproduces the issue. The expected locations of the STL shared objects
are part of the sanity check of the NDK, but we never added an NdkInfo
for r18 to account for the removal of the old STL flavors.

It looks like I also forgot to make use of the NdkR17Info that I
added, which means that any build using prefab libraries with the
default STL and with r17 or r18 is probably broken.

Test: built sample with r18
Bug: None
Change-Id: I21da92433ee63ab2da10fd144bd3dbb7cff2fa52