Factor interface from ExternalNativeJsonGenerator
For CMake vNext support we want to skip the intermediate phase of generating
android_gradle_build.json file and instead allow the system to directly
generate the Gradle model that Android Studio expects. This is so that we
can translate directly from compile_commands.json or CMake query directly
into the model.

This CL only factors out the interface, it doesn't yet change the interface
not depend on android_gradle_build.json semantics.

Bug: 153964094
Test: NA
Change-Id: Ia7b66c06f4380cee209637817c36c9f690233139