Improve performance of AndroidX dependency substitution
There are 2 ways of writing dependency substitution rules (see the
discussion at https://issuetracker.google.com/156449751):
  1. Simple and fast (https://docs.gradle.org/current/userguide/resolution_rules.html#sub:module_to_project_substitution)
  2. Complex and slow (https://docs.gradle.org/current/userguide/resolution_rules.html#sub:conditional_dependency_substitution)

The first one allows dependencies to be resolved during the task
execution phase, whereas the second one forces Gradle to resolve
dependencies during task graph creation.

Previously, we had to use the second to perform the substitution
correctly for data binding (see https://issuetracker.google.com/78202536).

With this CL, we now use the first way to write the rules, relying on a
tiny bit of heuristics to check if a configration is an annotation
processor classpath:

  private fun isAnnotationProcessClasspath(configuration: Configuration)
          : Boolean {
      return configuration.name.endsWith("AnnotationProcessorClasspath")
	   || configuration.name.startsWith("kapt")
  }

This CL also adds a check in JetifierTest to prevent dependencies being
resolved during task graph creation. We could consider applying this
check to more tests later.

Bug: 156449751
Test: Updated JetifierTest, existing DataBindingJetifierTest

Change-Id: I5347d2208b01bab2f2e0558830bb3561ff449e4f