Refactoring ArtifactDependencyGraph.
Move LibraryDependencyCacheBuildService.instantiateLibrary
into ArtifactHandler so that the same logic can be used
for both Level1 and Level2, and in the future v2, as well
as the task that creates the LintModelDependencies.

Level2 (and future v2) still has to go via the cache
in LibraryDependencyCacheBuildService but Level1 does not
so this is driven by a DependenciesConsumer that's also
customized based on the needs.

Bug: N/A
Test: existing tests
Change-Id: Ib48605fe272cba56e4c4eed6c93eafe12d9d8443