Add hint text for DuplicateRelativeFileException
Add a hint text for DuplicateRelativeFileException, if more than one files were found with the same OS independent path in the case that these duplicate files are all in the app or dynamic-feature module.
This hint message is added to advice adding the packagingOptions block if that can help manually resolve the DuplicateRelativeFileException.

Test: MergeJavaResourcesDelegateTest
Bug: 37069473
Change-Id: I036af50b373e39d1feeed71d6eb05e9b15efc924