Target MacOSX 10.13 when building Mac artifacts
IntelliJ system requirements says MacOS 10.13 is the currently minimum
supported version, so we need to make sure we are targetting it when
producing our native artifacts (like the Trace Processor Daemon), in
order to guarantee compatibility.

Otherwise, it will target the system that is building those artifacts,
which for us is mostly 10.15 and can break when those are run in older
Mac OSX versions.

Bug: 159653460
Test: existing
Change-Id: I526efc11860c26b6758cc24afe8603fcea395f5d