Don't cache for non-IWI install on Run
Non-IWI install does not write an OID into the code cache; thus, we do
not trust our own cache entries and end up verifying the entries with a
full dump anyways.

This CL removes the verification and just uses the dump directly for
situations in which we don't trust the cache, rather than caching and
verifying.

Bug: None
Test: Manual
Change-Id: I9ad8178ef8bf066d6b0bd37bf318c6a458ae1496